Update App.tsx
Updated the items.map code so we render items appropriately depending on if items with strikethrough are hidden or not. When isHidden is false we render everything, but when true we only render items where item.strike is false.

Added showAddItemForm function def and added state to show and hide the add item form.