Add FormProps.model and update App and AddItemForm
Added FormProps.model so we can pass an onsubmit handler to our form from App.

Renamed the addItem function in App to handleFormSubmit. Passed it as onSubmit to AddItemForm.

Updated AddItemForm. It uses the new FormProps type. It now has state for the for input. Added a handleSubmit which calls the onSubmit we passed from App.tsx.