Update show/close form functionality
Updated App.tsx. Finish the showAddItemForm function that toggles the isFormHidden state. Passed this function to AddItemForm component and updated the FormProps. Added showAddItemForm to the + button so the user can open show the form. Added the onClose function to the X button in AddItemForm so the user can close the form.